Transaksi feature complete

+++ database/migrations/2023_08_26_041251_create_transaksi_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->integer('jumlah');
             $table->double('panjang');
             $table->double('lebar');
-            $table->integer('total_harga');
+            $table->double('total_harga');
             $table->string('keterangan');
             $table->timestamps();
         });
